+++ CHANGELOG.md
@@ -26,6 +26,7 @@
 * [#1340](https://github.com/bbatsov/projectile/pull/1340): Fix remote projects being removed if TRAMP can't connect.
 * [#1655](https://github.com/bbatsov/projectile/pull/1655): Fix `projectile-replace-regexp` searching the wrong files when called with prefix arg.
 * [#1659](https://github.com/bbatsov/projectile/issues/1659): Fix `projectile-project-vcs` to work outside a project.
+* [#1637](https://github.com/bbatsov/projectile/pull/1661): Integrate with savehist-mode
 
 ## 2.3.0 (2020-11-27)